requirements.txt
asgiref==3.8.1
Django==5.2.1
django-crispy-forms==2.4
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.19.15
pillow==11.2.1
python-decouple==3.8
slugify==0.0.1
sqlparse==0.5.3
